Add PatrolRewardModel
