+++ nekoyume/Assets/_Scripts/UI/Widget/Popup/PatrolRewardPopup.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Nekoyume.Game.Controller;
 using Nekoyume.L10n;
-using Nekoyume.UI.Model;
+using Nekoyume.UI.Model.Patrol;
 using Nekoyume.UI.Module;
 using TMPro;
 using UnityEngine;
